+++ src/lib/scriptengine/generic_script_test.rs
@@ -7,6 +7,8 @@ fn execute_shell() {
         &vec!["exit 0".to_string()],
         test::get_os_runner(),
         test::get_os_extension(),
+        None,
+        &vec![],
         true,
     );
 }
@@ -18,6 +20,8 @@ fn execute_shell_error() {
         &vec!["exit 1".to_string()],
         test::get_os_runner(),
         test::get_os_extension(),
+        None,
+        &vec![],
         true,
     );
 }
@@ -28,6 +32,45 @@ fn execute_shell_error_no_validate() {
         &vec!["exit 1".to_string()],
         test::get_os_runner(),
         test::get_os_extension(),
+        None,
+        &vec![],
         false,
+    );
+}
+
+#[test]
+fn execute_shell_empty_arguments() {
+    execute(
+        &vec!["exit 0".to_string()],
+        test::get_os_runner(),
+        test::get_os_extension(),
+        Some(vec![]),
+        &vec![],
+        true,
+    );
+}
+
+#[test]
+fn execute_shell_cli_arguments() {
+    execute(
+        &vec!["exit $1".to_string()],
+        test::get_os_runner(),
+        test::get_os_extension(),
+        Some(vec![]),
+        &vec!["0".to_string()],
+        true,
+    );
+}
+
+#[test]
+#[should_panic]
+fn execute_shell_cli_arguments_error() {
+    execute(
+        &vec!["exit $1".to_string()],
+        test::get_os_runner(),
+        test::get_os_extension(),
+        Some(vec![]),
+        &vec!["1".to_string()],
+        true,
     );
 }