+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.16.7
 
+* shebang line support #130 (thanks @SamuelMarks)
 * Prevent multiple calls for init/end #184
 * Test internal makefile coverage tasks #183